Put batch splitting/backoff behind config flag
